+++ packages/plugin-gallery/web/src/getAddButtons.tsx
@@ -47,9 +47,8 @@ export const getAddButtons = (config, galleryPluginService): AddButton[] => {
       icon: InsertPluginIcon,
       tooltip: 'GalleryPlugin_InsertButton_Tooltip',
       toolbars: [TOOLBARS.MOBILE, TOOLBARS.FOOTER, TOOLBARS.SIDE],
-      command: (editorCommands, uploadContext) => {
-        if (uploadContext) {
-          const { uploadService, updateService } = uploadContext;
+      command: (editorCommands, uploadService, updateService) => {
+        if (uploadService && updateService) {
           if (config.handleFileSelection) {
             config.handleFileSelection(
               undefined,